+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is mostly based on [Keep a Changelog](https://keepachangelog.com/en/1
 You can also check [on GitHub](https://github.com/nextcloud/news/releases), the release notes there are generated automatically and include every pull request.
 
 # Unreleased
+### Changed
+
+### Fixed
+- global starred count not updated when deleting a feed with starred items
 
 # Releases
 ## [28.0.0-beta.2] - 2026-01-12